+++ src/StreamSocket.h
@@ -9,7 +9,7 @@ class StreamSocket {
 
 	public:
 		StreamSocket();
-		StreamSocket(unsigned short fileDescriptor);
+		StreamSocket(int fileDescriptor);
 		~StreamSocket();
 		operator int() const;
 		void throwIfError() const;
